Move classes that belong to task APIs into task package
