+++ src/main/java/org/gradlex/maven/plugin/development/MavenPluginDevelopmentPlugin.java
@@ -32,8 +32,8 @@ import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.SourceSetContainer;
 import org.gradle.api.tasks.TaskProvider;
 import org.gradle.jvm.tasks.Jar;
-import org.gradlex.maven.plugin.development.internal.GAV;
-import org.gradlex.maven.plugin.development.internal.MavenPluginDescriptor;
+import org.gradlex.maven.plugin.development.task.GAV;
+import org.gradlex.maven.plugin.development.task.MavenPluginDescriptor;
 import org.gradlex.maven.plugin.development.task.DependencyDescriptor;
 import org.gradlex.maven.plugin.development.task.GenerateHelpMojoSourcesTask;
 import org.gradlex.maven.plugin.development.task.GenerateMavenPluginDescriptorTask;
