+++ src/main/java/org/gradlex/maven/plugin/development/task/GenerateMavenPluginDescriptorTask.java
@@ -40,7 +40,6 @@ import org.gradle.api.tasks.OutputDirectory;
 import org.gradle.api.tasks.PathSensitive;
 import org.gradle.api.tasks.PathSensitivity;
 import org.gradle.api.tasks.TaskAction;
-import org.gradlex.maven.plugin.development.internal.GAV;
 import org.gradlex.maven.plugin.development.internal.MavenLoggerAdapter;
 import org.gradlex.maven.plugin.development.internal.MavenServiceFactory;
